# apps/console/public/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/{userId}": {
      "get": {
        "operationId": "UsersController_getUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/invitations": {
      "post": {
        "operationId": "UserInvitationsController_inviteUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "UserInvitations"
        ]
      },
      "get": {
        "operationId": "UserInvitationsController_getInvitations",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserInvitations"
        ]
      }
    },
    "/api/v1/invitations/by-token/{token}": {
      "get": {
        "operationId": "UserInvitationsController_getInvitationByToken",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserInvitations"
        ]
      }
    },
    "/api/v1/invitations/accept": {
      "post": {
        "operationId": "UserInvitationsController_acceptInvitation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "UserInvitations"
        ]
      }
    },
    "/api/v1/invitations/{invitationId}": {
      "delete": {
        "operationId": "UserInvitationsController_cancelInvitation",
        "parameters": [
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserInvitations"
        ]
      }
    },
...
    "/api/v1/feature-flags/{key}": {
      "get": {
        "operationId": "FeatureFlagsController_getFlagByKey",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feature flag retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureFlagDto"
                }
              }
            }
          },
          "404": {
            "description": "Feature flag not found"
          }
        },
        "summary": "Get specific feature flag by key",
        "tags": [
          "Feature Flags"
        ]
      },
      "put": {
        "operationId": "FeatureFlagsController_updateFlag",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeatureFlagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Feature flag updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureFlagDto"
                }
              }
            }
          }
        },
        "summary": "Update feature flag enabled status",
        "tags": [
          "Feature Flags"
        ]
      }
    },
    "/api/v1/feature-flags/{key}/enabled": {
      "get": {
        "operationId": "FeatureFlagsController_isEnabled",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feature enabled status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string"
                    },
                    "enabled": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Check if feature is enabled",
        "tags": [
          "Feature Flags"
        ]
      }
    },
...
  },
  "info": {
    "title": "SALLY API",
    "description": "Fleet Operations Assistant API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://sally-api.apps.appshore.in/api/v1",
      "description": "Staging"
    },
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "type": "http",
        "description": "Enter your API key (sk_staging_...)",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
      "InviteUserDto": {
        "type": "object",
        "properties": {}
      },
      "AcceptInvitationDto": {
        "type": "object",
        "properties": {}
      },
...
      "FeatureFlagDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "route_planning_enabled"
          },
          "name": {
            "type": "string",
            "example": "Route Planning"
          },
          "description": {
            "type": "string",
            "example": "Intelligent route planning with HOS compliance"
          },
          "enabled": {
            "type": "boolean",
            "example": false
          },
          "category": {
            "type": "string",
            "example": "dispatcher"
          }
        },
        "required": [
          "key",
          "name",
          "enabled",
          "category"
        ]
      },
...
      "UpdateFeatureFlagDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "example": true,
            "description": "Enable or disable the feature"
          }
        },
        "required": [
          "enabled"
        ]
      },
...
    }
  }
}